Factor out yaml logic
